Wait before swap, not after
This change resulted from a discussion with @fintelia about how to
reduce the time spent waiting for readers. His observation was that it'd
be better if waiting for readers to drain following a swap could be
overlapped with computation on the part of the writer. In essence, the
writer might as well do useful work while waiting for readers, rather
than block doing nothing. The previous implementation of `refresh()`
precluded this, as it did not return until all old readers had departed.

This patch modifies that behavior. `refresh()` will now *first* wait for
readers to train from the *previous* swap, *then* play back the
operational log, and *then* swap the atomic pointer just before
returning.

This gives the same semantics as the previous version (writes are only
made visible after a `refresh()`), but improves write throughput as the
writer can do useful work while the old readers are draining.