De-duplicate values between the two maps
This is a major change.
Previously, evmap would store two copies of the values in the map.
Now, it stores two aliases of that data instead, which reduces memory
usage for maps with large values to about half of what it used to be.
It should also not impact read or write performance.

However, this change comes at the cost of an API complication. Values
must now implement the new `ShallowCopy` trait to indicate that they can
be aliases. This trait is already implemented for common wrapper types
(Box, Vec, Arc, Rc) and Copy types (primitives, references, and tuples),
but may need to be implemented for any custom types.

This change also removes the need for values to implement Clone, which
may make some users happy.