Avoid deadlock when registering new readers
Previously we attempted to take the epochs lock while holding onto our
handle of Inner. This causes a deadlock when a writer is also trying to
refresh(): the writer is holding the epochs lock and waiting for the
readers epoch to tick, while the reader is withholding its epoch tick
while waiting for the epoch lock.

By instead just cloning the Arc while holding the Inner, and taking the
lock *after* we've ticked the epoch, this issue is fixed.