refactor: remove hashbrown dependency
Now that hashbrown has long since been upstreamed into std, it seems
reasonable to remove this dependency. This is technically a breaking
change because we no longer export hashbrown's DefaultHashBuilder (was
evmap::FxHashBuilder), which could have performance implications for
consumers who which to use hashbrown's builder (which may still be
faster than the one in libstd).