Move to a guard-based read API
Previously, most of the evmap APIs were closure based. Specifically, for
a method like `get`, you had to provide a `then` function that mapped a
reference to a value contained in the map into an owned value. This was
there to ensure that we always reset the parity bit of the epoch after
the read operation completed.

That API was fairly annoying to work with. This patch switches to a new,
guard-based API. Where you would previously write something like:

```rust
map.get_and("foo", |values| {
    /* do stuff with values */
})
```

You can now write:

```rust
let values = map.get("foo");
// do stuff with values
drop(values);
```

Because `get` returns a `ReadGuard`, which `Deref`s to the inner type.
As long as the guard lives, the reference is valid. This API _does_ mean
it's easier to shoot yourself in the foot though. As long as the guard
lives, writers cannot refresh the map, since the guard is still holding
back the next epoch. Make sure you don't do too much work while holding
a guard, and especially blocking work!

This patch also introduces the `read` method on `ReadHandle`. It returns
a `MapReadRef`, which is sort of like getting a guard to the map as a
whole. It allows you to make multiple lookups and the like without
having to increment the epoch each time. It also implements the `Index`
and `IntoIterator` (for `&`) traits.