hashbag for large bags
This avoids a huge performance collapse when there are many values for a
given key and the user wants to remove one. We do this by switching
large values over to a `hashbag` for backing storage. This change means
that users no longer get a `&[V]`, but instead a `&Values<V>`, and also
that we require `V: Hash + Eq`.

With this change, I also made `smallvec` non-optional.

I also replaced the awkward `mem::forget` code with `ManuallyDrop` which
seems like a much safer approach. The failure mode is now memory leaks
rather than double-frees, which seems like a huge win.