Re-use epoch slots when ReadHandles go away
Previously, when a `ReadHandle` was dropped, its epoch in the "master"
epoch list would remain. This is unfortunate if there is a lot of
`ReadHandle` churn, for two reasons. First, it causes memory use to keep
increasing over time. And second, it means that the writer has to check
an increasing number of forever unchanging epochs.

This patch moves to `slab` for the epoch store, and removes the epoch
for a `ReadHandle` when it is dropped.

This isn't _technically_ a memory leak, because the memory would be
reclaimed when the last handle was eventually dropped, but in practice
it lead to ever-growing memory use.

Fixes #53.