First draft of a sound ShallowCopy
This is trying to address the unsoundness that arises from the current
version of `ShallowCopy` (see #74). In the process, it also deals with
the fact that casting between `Inner<.., T, ..>` and `Inner<..,
ManuallyDrop<T>, ..>` is likely not sound
(https://github.com/rust-lang/unsafe-code-guidelines/issues/35#issuecomment-731656729).
It does not yet work.