Fix loom test
**This Commit**
Adds a `loom::thread::yield_now` to `publish` to let `loom` advance
threads correctly.
It also adjusts the `loom` test so read in the thread instead of write.
This is because, if the write handle is dropped before the reader reads
(because the thread takes ownership of the write handle and completes),
the reader will return `None` instead of the correct value.